Refactor: 사물함 대여,반납 중복 로직 수정 및 필요없는 중복 컴포넌트 삭제#51
